Step2:로또(자동) Lotto / LottoStore 리팩토링
